refactored to accomodate new properties of Manager
