+++ src/keri/app/directing.py
@@ -25,41 +25,6 @@ from ..db import dbing
 logger = help.ogler.getLogger()
 
 
-def setupController(name="who", sith=None, count=1, temp=False,
-                    remotePort=5621, localPort=5620):
-    """
-    Setup and return doers list to run controller
-    """
-    # setup habitat
-    hab = habbing.Habitat(name=name, isith=sith, icount=count, temp=temp)
-    logger.info("\nDirect Mode controller %s:\nNamed %s on TCP port %s to port %s.\n\n",
-                hab.pre, hab.name, localPort, remotePort)
-
-    # setup doers
-    ksDoer = keeping.KeeperDoer(keeper=hab.ks)  # doer do reopens if not opened and closes
-    dbDoer = basing.BaserDoer(baser=hab.db)  # doer do reopens if not opened and closes
-
-    # setup wirelog to create test vectors
-    path = os.path.dirname(__file__)
-    path = os.path.join(path, 'logs')
-
-    wl = wiring.WireLog(samed=True, filed=True, name=name, prefix='keri',
-                        reopen=True, headDirPath=path)
-    wireDoer = wiring.WireLogDoer(wl=wl)
-
-    client = clienting.Client(host='127.0.0.1', port=remotePort, wl=wl)
-    clientDoer = doing.ClientDoer(client=client)
-    director = Director(hab=hab, client=client, tock=0.125)
-    reactor = Reactor(hab=hab, client=client)
-
-    server = serving.Server(host="", port=localPort, wl=wl)
-    serverDoer = doing.ServerDoer(server=server)
-    directant = Directant(hab=hab, server=server)
-    # Reactants created on demand by directant
-
-    return [ksDoer, dbDoer, wireDoer, clientDoer, director, reactor, serverDoer, directant]
-
-
 class Director(doing.Doer):
     """
     Base class for Direct Mode KERI Controller Doer with habitat and TCP Client
